shaving off more rom/ram

+++ include/cottage.h
@@ -12,7 +12,7 @@ namespace trog {
 
             bn::optional<bn::sprite_ptr> _flames;
             bn::optional<bn::sprite_animate_action<4>> _burningflames;
-            int _time_burning;
+            unsigned short _time_burning;
         public:
             cottage(bn::fixed xcor, bn::fixed ycor, direction direction, bool has_treasure, bool burninated);
             virtual void update() final;
